% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Distributions the Data was Generated From.}
\format{A tibble with two variables:
\describe{
  \item{rdist}{Distribution given in a form interpretable by R.}
  \item{rpar}{Parameters given as a list.}
}}
\usage{
distributions
}
\description{
A tibble comprising the density in the form interpretable by R; e.g., lnorm
instead of log-normal.
}
\keyword{datasets}
