% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_sd.R
\name{effect_sd}
\alias{effect_sd}
\title{Calculate the standard deviation of all studies' effects.}
\usage{
effect_sd(centre, spread, n, centre_type = "mean", spread_type = "sd")
}
\arguments{
\item{centre}{A sample mean or a median.}

\item{spread}{The associated measure of spread for the sample mean: either
a sample sd, sample interquartile range, or sample range.}

\item{n}{The sample size.}

\item{centre_type}{Specify if the center is "mean" or "median".}

\item{spread_type}{Specify if the spread is reported as "sd", "iqr", or "range".}
}
\description{
Estimate the standard deviation of the effect, depending on how that effect is reported
(median or mean).
}
