% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metamean.R
\name{metamean}
\alias{metamean}
\title{Meta-Analysis of the (difference of) means}
\usage{
metamean(df, mean_method = "mln", nboot = 1000, ...)
}
\arguments{
\item{df}{data frame containing the study-specific summary data. For one-group studies, this data frame can contain the following columns:
\tabular{ll}{
\code{min.g1} \tab minimum value. \cr
\code{q1.g1} \tab first quartile. \cr
\code{med.g1} \tab median. \cr
\code{q3.g1} \tab third quartile. \cr
\code{max.g1} \tab maximum value. \cr
\code{n.g1} \tab sample size. \cr
\code{mean.g1} \tab sample mean. \cr
\code{sd.g1} \tab sample standard deviation. \cr}
For two group studies, this data frame can also contain the following columns for the summary data of the second group: \code{min.g2}, \code{q1.g2}, \code{med.g2}, \code{q3.g2}, \code{max.g2}, \code{n.g2}, \code{mean.g2}, and \code{sd.g2}.}

\item{mean_method}{character string specifying the approach used to estimate the study-specific means. The options are the following:
\tabular{ll}{
\code{qe} \tab Quantile Matching Estimation (McGrath et al. 20220). \cr
\code{bc} \tab Box-Cox (McGrath et al. 2020). \cr
\code{mln} \tab Method for Unknown Non-Normal Distributions (Cai et al. 2021). This is the default option. \cr}}

\item{nboot}{integer specifying the number of bootstrap samples to use when estimating the study-specific standard errors. The default is \code{1000}.}

\item{...}{optional arguments that are passed into the \code{\link[metafor]{rma.uni}} function for pooling. See documentation of \code{\link[metafor]{rma.uni}}.}
}
\value{
an object of class "rma.uni". See documentation of \code{\link[metafor]{rma.uni}}.
}
\description{
The function meta-analyzes one-group or two-group studies where each study reports one of the following summary measures: \itemize{
\item S1: median, minimum and maximum values, and sample size
\item S2: median, first and third quartiles, and sample size
\item S3: median, minimum and maximum values, first and third quartiles, and sample size
\item S4: mean, standard deivation, and sample size.
 }
This function estimates the study-specific means and their standard errors from the S1, S2, S3, or S4 summary data. Then, this function estimates the pooled mean (for one-group studies) or the pooled difference of means (for two-group studies) based on the standard inverse variance method via the \code{\link[metafor]{rma.uni}} function. The convention used for calculating differences of means in two-group studies is: mean in group 1 minus mean in group 2.
}
\details{
When studies report S1, S2, or S3 summary data, the approaches described by McGrath et al. (2020) and Cai et al. (2021) can be applied to estimate the study-specific means. Their standard errors are estimated by parametric bootstrap (McGrath et al. 2022).
}
\references{
McGrath S., Zhao X., Steele R., Thombs B.D., Benedetti A., and the DEPRESsion Screening Data (DEPRESSD) Collaboration. (2020). Estimating the sample mean and standard deviation from commonly reported quantiles in meta-analysis. \emph{Statistical Methods in Medical Research}. \strong{29}(9):2520-2537.

Cai S., Zhou J., and Pan J. (2021). Estimating the sample mean and standard deviation from order statistics and sample size in meta-analysis. \emph{Statistical Methods in Medical Research}. \strong{30}(12):2701-2719.

McGrath S., Katzenschlager S., Zimmer A.J., Seitel A., Steele R., Benedetti A. (2022). Standard error estimation in meta-analysis of studies reporting medians. \emph{arXiv e-prints}.
}
