\name{metamedian-defunct}
\alias{metamedian-defunct}
\title{Defunct functions in package \sQuote{metamedian}}

\description{
  These functions are defunct and no longer available.
}

\details{
  The following functions are defunct and no longer available:
  \itemize{
    \item{\code{qe.fit} calls should now be \code{\link[estmeansd]{qe.fit}} (in the \sQuote{estmeansd} package) calls. That is, this function has been moved to the \sQuote{estmeansd} pacakge.}
    \item{Similarly, \code{print.qe.fit} calls should now be \code{\link[estmeansd]{print.qe.fit}} (in the \sQuote{estmeansd} package) calls.}
  }
}
