\name{metamedian-deprecated}
\alias{metamedian-deprecated}
\title{Deprecated functions in package \sQuote{metamedian}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{metamedian} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{\code{\link{qe.fit}} calls should now be \code{\link[estmeansd]{qe.fit}} (in the \sQuote{estmeansd} package) calls. That is, this function has been moved to the \sQuote{estmeansd} pacakge.}
    \item{Similarly, \code{\link{print.qe.fit}} calls should now be \code{\link[estmeansd]{print.qe.fit}} (in the \sQuote{estmeansd} package) calls.}

  }
}
