% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_studies.R
\name{print.describe_studies}
\alias{print.describe_studies}
\title{Print method for "describe_studies" objects}
\usage{
\method{print}{describe_studies}(x, ...)
}
\arguments{
\item{x}{object of class "describe_studies".}

\item{...}{other arguments.}
}
\value{
No value is returned.
}
\description{
Print method for objects of class "describe_studies"
}
\seealso{
\code{\link{describe_studies}}
}
