% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.qe.fit.R
\name{print.qe.fit}
\alias{print.qe.fit}
\title{Print method for objects of class "qe.fit"}
\usage{
\method{print}{qe.fit}(x, ...)
}
\arguments{
\item{x}{object of class "qe.fit".}

\item{...}{other arguments.}
}
\value{
No value is returned.
}
\description{
\emph{Note}: As of release 0.1.2, this function is deprecated; \code{\link[estmeansd]{print.qe.fit}} (in the \sQuote{estmeansd} package) should be used instead.

Print method for objects of class "qe.fit".
}
\examples{
\donttest{
## Generate S2 summary data
set.seed(1)
n <- 100
x <- stats::rlnorm(n, 2.5, 1)
quants <- stats::quantile(x, probs = c(0.25, 0.5, 0.75))

## Fit distributions
res <- qe.fit(q1.val = quants[1], med.val = quants[2], q3.val = quants[3],
       n = n)
print(res)
}

}
\seealso{
\code{\link{metamedian-deprecated}}, \code{\link{qe.fit}}
}
