% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_par.R
\docType{data}
\name{ss_par}
\alias{ss_par}
\title{Parameters of the Simulation}
\format{A tibble
\describe{
  \item{distributions}{A list of lists containing distribution character string distribution and a vector of parameters.}
  \item{trials}{The number of trials per sample size and distribution.}
 }}
\usage{
ss_par
}
\description{
These paramaters generated the simulated data.
}
\keyword{datasets}
